% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_basedata.R
\name{get_basedata}
\alias{get_basedata}
\title{Load basing data}
\usage{
get_basedata(host = NA, country_count = FALSE)
}
\arguments{
\item{host}{The Correlates of War (COW) numeric country code for the host country or countries in the series}

\item{country_count}{Logical. Should the function return a country-level count of the total number of bases or the country-site data}
}
\value{
Returns a data frame containing information on bases present within selected host countries
}
\description{
Load basing data
}
