% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{troopdata}
\alias{troopdata}
\title{Troop Data}
\format{
A data frame with country year observations including the following variables:
\describe{
\item{\code{ccode}}{A numeric vector of Correlates of War country codes.}
\item{\code{year}}{The year of the observation.}
\item{\code{troops}}{The total number of US military personnel deployed to the host country.}
\item{\code{army}}{Total number of Army personnel deployed to the host country.}
\item{\code{navy}}{Total number of Navy personnel deployed to the host country.}
\item{\code{air_force}}{Total number of Air Force personnel deployed to the host country.}
\item{\code{marine_corps}}{Total number of Marine Corps personnel deployed to the host country.}
}
}
\source{
\url{https://www.heritage.org/defense/report/global-us-troop-deployment-1950-2005}
}
\usage{
troopdata
}
\value{
Data on US military deployments to overseas locations (countries and territories) from 1950 through 2020. The data were originally
compiled by Tim Kane of the Heritage Foundation using Defense Manpower Data Center quarterly reports on US troop deployments.
These data originally ended in 2005 but we have updated them to run through 2020. We have also used news reports to update deployment
figures for particular countries like Iraq where the DOD stopped reporting data.
}
\description{
\code{get_troopdata()} returns a data frame containing information on US military deployments.
}
\keyword{datasets}
