% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basedata}
\alias{basedata}
\title{Base data}
\format{
A data frame with country-base or country observations including the following variables:
\describe{
\item{\code{countryname}}{A character vector of countrynames.}
\item{\code{ccode}}{A numeric vector of Correlates of War country codes.}
\item{\code{iso3c}}{A character vector of ISO three character country codes.}
\item{\code{basename}}{Name of the facility.}
\item{\code{lat}}{The facility's latitude.}
\item{\code{lon}}{The facility's longitude.}
\item{\code{base}}{Binary indicator identifying the facility as a major base or not.}
\item{\code{lilypad}}{A binary indicator identifying the facility as a lilypad or not. Vine codes lilypads as less than 200 personnel or "other site" designation in Pentagon reports.}
\item{\code{fundedsite}}{A binary variable indicating whether or not the facility is a host-state base funded by the US.}
}
}
\source{
\url{https://dra.american.edu/islandora/object/auislandora\%3A81234}
}
\usage{
basedata
}
\value{
Data on US military bases from the Cold War period through the present. These data were initially collected by David Vine (2015). We
updated the data through 2018 for inclusion in this package. The data include two versions. The first is the country-base data, which
also includes information on the type of military facility and the longitude and latitude of the facility. The second version presents
a simple country-count data frame, where the count is the number of bases located in that country.
}
\description{
\code{get_basedata()} returns a data frame containing David Vine's US basing data.
}
\keyword{datasets}
