% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basedata}
\alias{basedata}
\title{Data on US military bases from the Cold War period through the present. These data were initially collected by David Vine (2015). We
updated the data through 2018 for inclusion in this package. The data include two versions. The first is the country-base data, which
also includes information on the type of military facility and the longitude and latitude of the facility. The second version presents
a simple country-count data frame, where the count is the number of bases located in that country.}
\format{
A data frame with country year observations including the following variables:
\describe{
\item{\code{country}}{A character string giving the host country's name.}
\item{\code{ccode}}{A numeric vector of Correlates of War country codes.}
\item{\code{basename}}{Name of the facility.}
\item{\code{lat}}{The facility's latitude.}
\item{\code{lon}}{The facility's longitude.}
\item{\code{base}}{Binary indicator identifying the faciltiy as a major base  or not.}
\item{\code{lilypad}}{A binary indicator identifying the faciltiy as a lilypad or not.}
\item{\code{fundedsite}}{A binary variable indicating whether or not the facility is }
}
}
\source{
\url{https://dra.american.edu/islandora/object/auislandora\%3A81234}
}
\usage{
basedata
}
\description{
Data on US military bases from the Cold War period through the present. These data were initially collected by David Vine (2015). We
updated the data through 2018 for inclusion in this package. The data include two versions. The first is the country-base data, which
also includes information on the type of military facility and the longitude and latitude of the facility. The second version presents
a simple country-count data frame, where the count is the number of bases located in that country.
}
\keyword{datasets}
