% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_troopdata.R
\name{get_troopdata}
\alias{get_troopdata}
\title{Load Troop Data}
\usage{
get_troopdata(host = NA, branch = FALSE, startyear, endyear)
}
\arguments{
\item{host}{The Correlates of War (COW) numeric country code for the host country or countries in the series}

\item{startyear}{the first year for the series}

\item{endyear}{The last year for the series}
}
\value{
Returns a data frame containing country  year observations for troop deployments
}
\description{
Load Troop Data
}
