% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{troopdata}
\alias{troopdata}
\title{U.S. overseas troop deployment data}
\format{
A data frame with country year observations including the following variables:

\describe{
\item{\code{countryname}}{A character vector of country names.}
\item{\code{ccode}}{A numeric vector of Correlates of War country codes.}
\item{\code{iso3c}}{A character vector of ISO three character country codes.}
\item{\code{region}}{Correlates of War geographic region name.}
\item{\code{year}}{The year of the observation.}
\item{\code{troops}}{The total number of US military personnel deployed to the host country.}
\item{\code{army}}{Total number of Army personnel deployed to the host country.}
\item{\code{navy}}{Total number of Navy personnel deployed to the host country.}
\item{\code{air_force}}{Total number of Air Force personnel deployed to the host country.}
\item{\code{marine_corps}}{Total number of Marine Corps personnel deployed to the host country.}
}
}
\source{
\url{https://www.heritage.org/defense/report/global-us-troop-deployment-1950-2005}
}
\usage{
troopdata
}
\value{
Returns the full data frame containing country-year observations of US military deployments to overseas locations (countries and territories) from 1950 through 2020.
}
\description{
\code{troopdata} returns a data frame containing information on US military deployments.
}
\keyword{datasets}
