% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{troopdata_rebuild_reports}
\alias{troopdata_rebuild_reports}
\title{DMDC Deployment Reports}
\format{
A data frame with country year quarter observations including the following variables:

\describe{
\item{\code{ccode}}{A numeric vector of Correlates of War country codes.}
\item{\code{iso3c}}{A character vector of ISO three character country codes.}
\item{\code{countryname}}{A character vector of country names.}
\item{\code{region}}{Correlates of War geographic region name.}
\item{\code{year}}{The year of the observation.}
\item{\code{month}}{The month of the observation.}
\item{\code{quarter}}{The quarter of the observation.}
\item{\code{source}}{The DMDC report source of the observation.}
\item{\code{Location}}{The geographic location listed in the DMDC reports.}
\item{\code{Total}}{"Total number of US military personnel deployed to the host country.}
\item{\code{Total Ashore}}{"Total number of US military personnel deployed to the host country, excluding those at sea.}
\item{\code{Total Afloat}}{"Total number of US military personnel deployed to the host country, at sea.}
\item{\code{Army Total}}{Total number of Army personnel deployed to the host country.}
\item{\code{Navy Ashore}}{Total number of Navy personnel deployed to the host country, excluding those at sea.}
\item{\code{Navy Temporary Ashore}}{Total number of Navy personnel deployed to the host country, temporarily.}
\item{\code{Navy Other}}{Total number of Navy personnel deployed to the host country, in other capacities.}
\item{\code{Marine Corps Ashore}}{Total number of Marine Corps personnel deployed to the host country, excluding those at sea.}
\item{\code{Marine Corps Afloat}}{Total number of Marine Corps personnel deployed to the host country, at sea.}
\item{\code{Air Force Total}}{Total number of Air Force personnel deployed to the host country.}
\item{\code{Navy Afloat}}{Total number of Navy personnel deployed to the host country, at sea.}
\item{\code{Navy Total}}{Total number of Navy personnel deployed to the host country.}
\item{\code{Marine Corps Total}}{Total number of Marine Corps personnel deployed to the host country.}
\item{\code{troops_ad}}{The total number of active duty US military personnel deployed to the host country.}
\item{\code{army_ad}}{Total number of active duty Army personnel deployed to the host country.}
\item{\code{navy_ad}}{Total number of active duty Navy personnel deployed to the host country.}
\item{\code{marine_corps_ad}}{Total number of active duty Marine Corps personnel deployed to the host country.}
\item{\code{space_force_ad}}{Total number of active duty Space Force personnel deployed to the host country.}
\item{\code{air_force_ad}}{Total number of active duty Air Force personnel deployed to the host country.}
\item{\code{coast_guard_ad}}{Total number of Coast Guard personnel deployed to the host country.}
\item{\code{Macro Location}}{The geographic location listed in the DMDC reports.}
\item{\code{Army Active Duty}}{Total number of active duty Army personnel deployed to the host country.}
\item{\code{Navy Active Duty}}{Total number of active duty Navy personnel deployed to the host country.}
\item{\code{Marine Corps Active Duty}}{Total number of active duty Marine Corps personnel deployed to the host country.}
\item{\code{Air Force Active Duty}}{Total number of active duty Air Force personnel deployed to the host country.}
\item{\code{Coast Guard Active Duty}}{Total number of active duty Coast Guard personnel deployed to the host country.}
\item{\code{Space Force Active Duty}}{Total number of active duty Space Force personnel deployed to the host country.}
\item{\code{Total Active Duty}}{Total number of active duty US military personnel deployed to the host country.}
\item{\code{Army National Guard}}{Total number of Army National Guard personnel deployed to the host country.}
\item{\code{Army Reserve}}{Total number of reserve Army personnel deployed to the host country.}
\item{\code{Navy Reserve}}{Total number of reserve Navy personnel deployed to the host country.}
\item{\code{Marine Corps Reserve}}{Total number of reserve Marine Corps personnel deployed to the host country.}
\item{\code{Air National Guard}}{Total number of Air National Guard personnel deployed to the host country.}
\item{\code{Air Force Reserve}}{Total number of reserve Air Force personnel deployed to the host country.}
\item{\code{Coast Guard Reserve}}{Total number of reserve Coast Guard personnel deployed to the host country.}
\item{\code{Total Selected Reserve}}{Total number of reserve US military personnel deployed to the host country.}
\item{\code{Army Civilian}}{Total number of Army civilian personnel deployed to the host country.}
\item{\code{Navy Civilian}}{Total number of Navy civilian personnel deployed to the host country.}
\item{\code{Marine Corps Civilian}}{Total number of Marine Corps civilian personnel deployed to the host country.}
\item{\code{Air Force Civilian}}{Total number of Air Force civilian personnel deployed to the host country.}
\item{\code{DOD Civilian}}{Total number of Department of Defense civilian personnel deployed to the host country.}
\item{\code{Total Civilian}}{Total number of civilian personnel deployed to the host country.}
\item{\code{Grand Total}}{Total number of US military and civilian personnel deployed to the host country.}
}
}
\source{
\url{https://www.heritage.org/defense/report/global-us-troop-deployment-1950-2005}

\doi{10.1177/07388942211030885}
}
\usage{
troopdata_rebuild_reports
}
\value{
Returns a data frame containing DMDC reports of US military deployments to overseas locations from 1950 through 2024.
}
\description{
\code{troopdata_rebuild_reports} returns a data frame containing DMDC reports on US military deployments.
}
\keyword{datasets}
